Fixed Scopus issues
Changed way columns to parse are defined.  No longer hard coded by position.  Looks in the column header to figure out which we need.  Will work for older scopus files as well as new ones (and it allows users to download whatever file they like as long as they have the essential information)